requirements.txt
aws-requests-auth>=0.3.0
APScheduler>=3.5.1
blist>=1.3.6
boto3>=1.4.4
configparser>=3.5.0
croniter>=0.3.16
cryptography>=2.1.4
elasticsearch==7.0.0
envparse>=0.2.0
exotel>=0.1.3
jira>=1.0.10,<1.0.15
jsonschema>=2.6.0
mock>=2.0.0
pymongo==3.6.1
PyStaticConfiguration>=0.10.3
python-dateutil>=2.6.0,<2.7.0
PyYAML>=3.12
requests>=2.0.0
stomp.py>=4.1.17
texttable>=0.8.8
twilio==6.0.0
six>=1.11.0
